+++ assets/css/styles.css
@@ -1,0 +1,57 @@
+body {
+    font-family: 'Segoe UI', sans-serif;
+}
+
+.block-container {
+    max-width: 900px;
+    margin: 0 auto;
+    padding: 2rem;
+}
+
+.project-header {
+    text-align: center;
+    margin-top: 1rem;
+    margin-bottom: 2rem;
+}
+
+.home-container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 70vh;
+}
+
+.home-card {
+    background: #ffffff;
+    border-radius: 12px;
+    padding: 2rem;
+    box-shadow: 0 8px 20px rgba(0, 0, 0, 0.1);
+    max-width: 600px;
+    text-align: center;
+}
+
+.about-box {
+    background-color: #f1f3f6;
+    border-left: 5px solid #4a90e2;
+    padding: 1rem;
+    margin-bottom: 1.5rem;
+    border-radius: 6px;
+    font-size: 0.95rem;
+}
+
+img {
+    border-radius: 10px;
+}
+
+/* Reduce sidebar width */
+.css-1d391kg, .css-1d391kg > div {
+    width: 250px !important;
+}
+
+/* Standard text sizes */
+h1 { font-size: 2.2rem; }
+h2 { font-size: 1.5rem; }
+p, li { font-size: 1rem; }
+
+/* Sidebar tweaks */
+.css-1lcbmhc { padding-top: 2rem; }
