+++ config.py
@@ -1,0 +1,15 @@
+CLASS_LABELS = ['onion', 'pear', 'strawberry', 'tomato']
+
+MODEL_PATH = "assets/modelWeights/best_model_v1.pth"
+MODEL_PATH_ONION = "assets/modelWeights/best_model_onion_v1.pth"
+MODEL_PATH_PEAR  = "assets/modelWeights/best_model_pear_v1.pth"
+MODEL_PATH_TOMATO  = "assets/modelWeights/best_model_tomato_v1.pth"
+MODEL_PATH_STRAWBERRY  = "assets/modelWeights/best_model_strawberry_v1.pth"
+
+GOOGLE_DRIVE_FILES = {
+    'assets/nlp/torch_recipe_embeddings_231630.pt': '1PSidY1toSfgECXDxa4pGza56Jq6vOq6t',
+    'assets/nlp/tag_based_bert_model.pth': '1LBl7yFs5JFqOsgfn88BF9g83W9mxiBm6',
+    'assets/nlp/RAW_recipes.csv': '1rFJQzg_ErwEpN6WmhQ4jRyiXv6JCINyf',
+    'assets/nlp/recipe_statistics_231630.pkl': '1n8TNT-6EA_usv59CCCU1IXqtuM7i084E',
+    'assets/nlp/recipe_scores_231630.pkl': '1gfPBzghKHOZqgJu4VE9NkandAd6FGjrA'
+}
