--- packages.txt
@@ -1,2 +1,0 @@
-build-essential
-python3-dev
