+++ requirements.txt
@@ -1,9 +1,56 @@
-torch==2.5.0
-transformers==4.51.1
-tokenizers==0.21.1
-streamlit==1.44.1
-pandas==2.2.3
-toml==0.10.2
-numpy==2.0.2
-requests==2.32.3
+altair==5.5.0
+attrs==25.3.0
+beautifulsoup4==4.13.4
+blinker==1.9.0
+cachetools==6.1.0
+certifi==2025.7.14
+charset-normalizer==3.4.2
+click==8.2.1
+filelock==3.18.0
+fsspec==2025.7.0
 gdown==5.2.0
+gitdb==4.0.12
+GitPython==3.1.44
+hf-xet==1.1.5
+huggingface-hub==0.33.4
+idna==3.10
+Jinja2==3.1.6
+jsonschema==4.24.1
+jsonschema-specifications==2025.4.1
+MarkupSafe==3.0.2
+mpmath==1.3.0
+narwhals==1.47.1
+networkx==3.5
+numpy==2.3.1
+packaging==25.0
+pandas==2.3.1
+pillow==11.3.0
+protobuf==6.31.1
+pyarrow==21.0.0
+pydeck==0.9.1
+PySocks==1.7.1
+python-dateutil==2.9.0.post0
+pytz==2025.2
+PyYAML==6.0.2
+referencing==0.36.2
+regex==2024.11.6
+requests==2.32.4
+rpds-py==0.26.0
+safetensors==0.5.3
+setuptools==80.9.0
+six==1.17.0
+smmap==5.0.2
+soupsieve==2.7
+streamlit==1.47.0
+sympy==1.14.0
+tenacity==9.1.2
+tokenizers==0.21.2
+toml==0.10.2
+torch==2.7.1
+torchvision==0.22.1
+tornado==6.5.1
+tqdm==4.67.1
+transformers==4.53.2
+typing_extensions==4.14.1
+tzdata==2025.2
+urllib3==2.5.0
